milestone 19

+++ frontend/src/components/CartCard.jsx
@@ -1,57 +1,42 @@
-import React from 'react'
-import styles from './card.module.css'
-import axios from 'axios'
-const cartCard = ({products}) => {
-
-  async function handleDelete(){
+import React, { useState } from 'react'
+import styles from "./card.module.css";
+import axios from "axios";
+const MyProductCard = ({product}) => {
+    const [quantity,setQuantity] = useState(1);
+  async function handleDelete(id){
     try {
+      const token = JSON.parse(localStorage.getItem("follow-along-auth-token-user-name-id"));
+      if (!token) {
+          alert("Please login first");
+          return;
+      }
+      console.log(token,id)
       await axios.delete(`http://localhost:8080/product/delete/${id}`,{
-        headers:{
-          Authorization: token.token
+        headers: { 
+            "Authorization": token.token 
         }
-      })
+    });
     } catch (error) {
-      alert("Something went wrong");
-      console.log(error);
+      alert("something went wrong");
+      console.log(error)
     }
   }
   return (
-    <div className={styles.Card}>
-        <img className={styles.productImg} src={products.images[0]} alt={products.title} />
-        <h3>{products.title}</h3>
-        <p>${products.price}</p>
-        <div
-        style={{
-          display:"flex",
-          justifyContent:"space-around",
-          padding:"0.5rem",
-        }}
-        >
-            <button className="btn-del-edt" style={{
-              backgroundColor:"gray",
-              border:"1px solid",
-              borderRadius:"0.3rem",
-              width:"3rem",
-              display:"flex",
-              justifyContent:"space-around",
-              margin:"auto"
-            }}>Edit</button>
-            <button className="btn-del-edt" style={{
-              backgroundColor:"gray",
-              border:"1px solid",
-              borderRadius:"0.3rem",
-              width:"3rem",
-              display:"flex",
-              justifyContent:"space-around",
-              margin:"auto"
-            }}
-            onClick={()=>handleDelete(products._id)}
-            >Delete</button>
+<div className={styles.card}>
+        <img className={styles.productImg} src={product.images[0]} alt={product.title} />
+        <h3>{product.title}</h3>
+        <p>${product.price}</p>
+        <div>
+          <button
+          onClick={()=>setQuantity(quantity-1)}
+          >-</button>
+          <button>{quantity}</button>
+          <button
+           onClick={()=>setQuantity(quantity+1)}
+          >+</button>
         </div>
-
-      
     </div>
   )
 }
 
-export default cartCard;
+export default MyProductCard;
